requirements_new.txt
flask
gunicorn
gspread
google-auth
google-api-python-client
openpyxl
reportlab
